Move boolean gate to its own action
