+++ lib/flipper/ui/middleware.rb
@@ -49,6 +49,7 @@ module Flipper
         @action_collection.add UI::Actions::Feature
         @action_collection.add UI::Actions::AddActor
         @action_collection.add UI::Actions::AddGroup
+        @action_collection.add UI::Actions::BooleanGate
         @action_collection.add UI::Actions::Gate
 
         # Catch all, always last.
