+++ spec/flipper/ui_spec.rb
@@ -2,7 +2,6 @@ require 'helper'
 require 'rack/test'
 require 'flipper'
 require 'flipper/adapters/memory'
-require 'flipper/ui/actor'
 
 describe Flipper::UI do
   include Rack::Test::Methods
@@ -11,12 +10,25 @@ describe Flipper::UI do
   let(:flipper) { Flipper.new(adapter) }
   let(:app)     { described_class.app(flipper) }
 
+  describe "Initializing middleware with flipper instance" do
+    let(:app) { described_class.app(flipper) }
+
+    it "works" do
+      flipper.enable :some_great_feature
+      get "/features"
+      last_response.status.should be(200)
+      last_response.body.should include("some_great_feature")
+    end
+  end
+
   describe "Initializing middleware lazily with a block" do
     let(:app) { described_class.app(lambda { flipper }) }
 
     it "works" do
+      flipper.enable :some_great_feature
       get "/features"
       last_response.status.should be(200)
+      last_response.body.should include("some_great_feature")
     end
   end
 end
